Validation bug resolved
If the parent attribute is required, has children, but does not have multiple as true, it didn't fit into validation.

ResourceType was changed to REQUIRED for testing purpose only and it failed validation.

This was resolved by checking for parentRequired and also by removing attributes with dot representation for which new array of objects gets created for multiple entries {here date.dateType, etc.}.